group unit test classes from async_emon_api module
